all the things except email bc no
