+++ browser/react/components/Main.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import AllAlbums from './AllAlbums';
+import StatefulAlbums from './StatefulAlbums';
 import SingleAlbum from './SingleAlbum';
 import Sidebar from './Sidebar';
 import Player from './Player';
@@ -8,39 +8,17 @@ import SingleArtist from './SingleArtist';
 import {
   HashRouter,
   Route,
+  Switch
 } from 'react-router-dom';
 
+
+const NoMatch = ({ location }) => (
+  <div>
+    <h3>NOOOOOOOOOoooooo No match for <code>{location.pathname}</code></h3>
+  </div>
+)
+
 export default class Main extends Component {
-
-  // constructor (props) {
-  //   super(props);
-  //   this.state = {
-  //     albums: [],
-  //     selectedAlbum: {}
-  //   };
-  //   this.selectAlbum = this.selectAlbum.bind(this);
-  //   this.deselectAlbum = this.deselectAlbum.bind(this);
-  // }
-
-  // componentDidMount () {
-  //   axios.get('/api/albums/')
-  //     .then(res => res.data)
-  //     .then(albums => {
-  //       this.setState({ albums })
-  //     });
-  // }
-
-  // selectAlbum (albumId) {
-  //   axios.get(`/api/albums/${albumId}`)
-  //     .then(res => res.data)
-  //     .then(album => this.setState({
-  //       selectedAlbum: album
-  //     }));
-  // }
-
-  // deselectAlbum () {
-  //   this.setState({ selectedAlbum: {}});
-  // }
 
   render () {
     return (
@@ -50,11 +28,14 @@ export default class Main extends Component {
           <Sidebar deselectAlbum={this.deselectAlbum} />
         </div>
           <div className="col-xs-10">
-          <Route exact path='/' component={AllAlbums} />
-          <Route exact path='/albums' component={AllAlbums} />
-          <Route path='/albums/:albumId' component={SingleAlbum} />
-          <Route exact path='/artists' component={AllArtists} />
-          <Route path='/artists/:artistId' component={SingleArtist} />
+            <Switch>
+              <Route exact path="/" component={StatefulAlbums} />
+              <Route exact path="/albums" component={StatefulAlbums} />
+              <Route path="/albums/:albumId" component={SingleAlbum} />
+              <Route exact path="/artists" component={AllArtists} />
+              <Route path="/artists/:artistId" component={SingleArtist} />
+              <Route component={NoMatch} />
+            </Switch>
         </div>
         <Player />
       </div>
@@ -63,9 +44,3 @@ export default class Main extends Component {
   }
 }
 
-// Take out the ternary expression that we're using to conditionally render AllAlbums and SingleAlbum:
-// {
-//   this.state.selectedAlbum.id ?
-//   <SingleAlbum album={this.state.selectedAlbum} /> :
-//   <AllAlbums albums={this.state.albums} selectAlbum={this.selectAlbum} />
-// }
