+++ browser/react/components/SingleArtist.js
@@ -1,62 +1,44 @@
 import React, { Component } from 'react';
 import Songs from '../components/Songs';
 import axios from 'axios';
+import AllAlbums from './AllAlbums';
+import {Route, NavLink} from 'react-router-dom';
 
 export default class SingleArtist extends Component {
 
       constructor () {
         super();
         this.state = {
-        artist: {},
-        artistAlbums: [],
-        artistSongs: []
+          artist: {},
+          artistAlbums: [],
+          artistSongs: []
         };
       }
 
       componentDidMount () {
         const artistId = this.props.match.params.artistId;
         Promise.all([
-            axios.get(`/api/artists/${artistId}`)
-             // .then(res => res.data)
-            //   .then(artist => {
-            //     this.setState({artist: artist });
-            //   })
-             ,
-              axios.get(`/api/artists/${artistId}/albums`)
-             // .then(res => res.data)
-            //   .then(albums => {
-            //     this.setState({artistAlbums: albums });
-            //   })
-              ,
-             axios.get(`/api/artists/${artistId}/songs`)
-              //.then(res => res.data)
-            //   .then(songs => {
-            //     this.setState({artistSongs: songs });
-             // })
+            axios.get(`/api/artists/${artistId}`),
+            axios.get(`/api/artists/${artistId}/albums`),
+            axios.get(`/api/artists/${artistId}/songs`)
         ])
-        //.then(res => res.data)
-        .then(data => {
-            console.log(data);
+        .then(([artist, albumsObj, songsObj]) => {
+            this.setState({ artist: artist.data, artistAlbums: albumsObj.data, artistSongs: songsObj.data });
         })
-
-
       }
       render () {
-        const artist = this.state.artist;
-        return (
+          const artist = this.state.artist; // or however you've named it
+          return (
             <div>
-            <h3>{artist.name}</h3>
-            <h4>ALBUMS</h4>
-            <h4>SONGS</h4>
+              <h3>{ artist.name }</h3>
+              <ul className="nav nav-tabs">
+                <li><NavLink to="/artists/:artistId/albums" activeClassName="selected">ALBUMS</NavLink></li>
+                <li><NavLink to="/artists/:artistId/songs" activeClassName="selected">SONGS</NavLink></li>
+              </ul>
+              <Route path="/artists/:artistId/albums" render={() => <AllAlbums albums={this.state.artistAlbums} />} />
+              <Route path="/artists/:artistId/songs" render={() => <Songs songs={this.state.artistSongs} />} />
             </div>
-        //   <div className="artist">
-        //     <div>
-        //       <h3>{ artist.name }</h3>
-        //       <img src={ artist.imageUrl } className="img-thumbnail" />
-        //     </div>
-        //     <Songs songs={artist.songs} />
-        //   </div>
-        );
-      }
+          );
+        }
     }
