+++ browser/react/components/StatefulAlbums.js
@@ -1,0 +1,85 @@
+import React, { Component } from 'react';
+import axios from 'axios';
+import {Link} from 'react-router-dom';
+import AllAlbums from './AllAlbums';
+
+const fakeAlbums = [
+  {
+    name: 'Abbey Road',
+    id: 1,
+    imageUrl: 'http://fillmurray.com/300/300',
+    songs: [
+      {
+        id: 1,
+        name: 'Romeo & Juliette',
+        artists: [
+          { name: 'Bill' }
+        ],
+        genre: 'Funk',
+        audioUrl: 'https://learndotresources.s3.amazonaws.com/workshop/5616dbe5a561920300b10cd7/Dexter_Britain_-_03_-_The_Stars_Are_Out_Interlude.mp3'
+      },
+      {
+        id: 2,
+        name: 'White Rabbit',
+        artists: [
+          { name: 'Bill' },
+          { name: 'Alice' }
+        ],
+        genre: 'Fantasy',
+        audioUrl: 'https://learndotresources.s3.amazonaws.com/workshop/5616dbe5a561920300b10cd7/Dexter_Britain_-_03_-_The_Stars_Are_Out_Interlude.mp3'
+      },
+      {
+        id: 3,
+        name: 'Lucy in the Sky with Diamonds',
+        artists: [
+          { name: 'Bob' }
+        ],
+        genre: 'Space',
+        audioUrl: 'https://learndotresources.s3.amazonaws.com/workshop/5616dbe5a561920300b10cd7/Dexter_Britain_-_03_-_The_Stars_Are_Out_Interlude.mp3'
+      }
+    ]
+  },
+  {
+    name: 'Yellow Submarine',
+    id: 2,
+    imageUrl: 'http://fillmurray.com/300/300',
+    songs: [
+      {
+        id: 4,
+        name: 'London Calling',
+        artists: [
+          { name: 'Bill' }
+        ],
+        genre: 'Punk',
+        audioUrl: 'https://learndotresources.s3.amazonaws.com/workshop/5616dbe5a561920300b10cd7/Dexter_Britain_-_03_-_The_Stars_Are_Out_Interlude.mp3'
+      }
+    ]
+  }
+];
+
+export default class StatefulAlbums extends Component {
+
+constructor (props) {
+  super(props);
+  this.state = {
+    albums: []
+  };
+}
+
+componentDidMount () {
+  axios.get('/api/albums/')
+    .then(res => res.data)
+    .then(albums => {
+      this.setState({ albums });
+    });
+}
+
+  render () {
+    const albums = this.state.albums;
+    return (
+      <div>
+          <AllAlbums albums={this.state.albums} />
+      </div>
+    );
+  }
+}
